+++ src/cryptography/hazmat/backends/openssl/backend.py
@@ -4,7 +4,6 @@
 
 from __future__ import absolute_import, division, print_function
 
-import base64
 import collections
 import contextlib
 import itertools
@@ -1807,6 +1806,16 @@ class Backend(object):
                 "Unsupported encoding for TraditionalOpenSSL"
             )
 
+        # OpenSSH + PEM
+        if format is serialization.PrivateFormat.OpenSSH:
+            if encoding is serialization.Encoding.PEM:
+                return ssh.serialize_ssh_private_key(key, password)
+
+            raise ValueError(
+                "OpenSSH private key format can only be used"
+                " with PEM encoding"
+            )
+
         # Anything that key-specific code was supposed to handle earlier,
         # like Raw.
         raise ValueError("format is invalid with this key")
@@ -1874,7 +1883,7 @@ class Backend(object):
         # OpenSSH + OpenSSH
         if format is serialization.PublicFormat.OpenSSH:
             if encoding is serialization.Encoding.OpenSSH:
-                return self._openssh_public_key_bytes(key)
+                return ssh.serialize_ssh_public_key(key)
 
             raise ValueError(
                 "OpenSSH format must be used with OpenSSH encoding"
@@ -1883,59 +1892,6 @@ class Backend(object):
         # Anything that key-specific code was supposed to handle earlier,
         # like Raw, CompressedPoint, UncompressedPoint
         raise ValueError("format is invalid with this key")
-
-    def _openssh_public_key_bytes(self, key):
-        if isinstance(key, rsa.RSAPublicKey):
-            public_numbers = key.public_numbers()
-            return b"ssh-rsa " + base64.b64encode(
-                ssh._ssh_write_string(b"ssh-rsa") +
-                ssh._ssh_write_mpint(public_numbers.e) +
-                ssh._ssh_write_mpint(public_numbers.n)
-            )
-        elif isinstance(key, dsa.DSAPublicKey):
-            public_numbers = key.public_numbers()
-            parameter_numbers = public_numbers.parameter_numbers
-            return b"ssh-dss " + base64.b64encode(
-                ssh._ssh_write_string(b"ssh-dss") +
-                ssh._ssh_write_mpint(parameter_numbers.p) +
-                ssh._ssh_write_mpint(parameter_numbers.q) +
-                ssh._ssh_write_mpint(parameter_numbers.g) +
-                ssh._ssh_write_mpint(public_numbers.y)
-            )
-        elif isinstance(key, ed25519.Ed25519PublicKey):
-            raw_bytes = key.public_bytes(serialization.Encoding.Raw,
-                                         serialization.PublicFormat.Raw)
-            return b"ssh-ed25519 " + base64.b64encode(
-                ssh._ssh_write_string(b"ssh-ed25519") +
-                ssh._ssh_write_string(raw_bytes)
-            )
-        elif isinstance(key, ec.EllipticCurvePublicKey):
-            public_numbers = key.public_numbers()
-            try:
-                curve_name = {
-                    ec.SECP256R1: b"nistp256",
-                    ec.SECP384R1: b"nistp384",
-                    ec.SECP521R1: b"nistp521",
-                }[type(public_numbers.curve)]
-            except KeyError:
-                raise ValueError(
-                    "Only SECP256R1, SECP384R1, and SECP521R1 curves are "
-                    "supported by the SSH public key format"
-                )
-
-            point = key.public_bytes(
-                serialization.Encoding.X962,
-                serialization.PublicFormat.UncompressedPoint
-            )
-            return b"ecdsa-sha2-" + curve_name + b" " + base64.b64encode(
-                ssh._ssh_write_string(b"ecdsa-sha2-" + curve_name) +
-                ssh._ssh_write_string(curve_name) +
-                ssh._ssh_write_string(point)
-            )
-        else:
-            raise ValueError(
-                "OpenSSH encoding is not supported for this key type"
-            )
 
     def _parameter_bytes(self, encoding, format, cdata):
         if encoding is serialization.Encoding.OpenSSH: